+++ tests/js/realm-tests.js
@@ -1465,4 +1465,31 @@ module.exports = {
         realm.close();
     },
 
+
+    testQueryBasedOnlyMethods: function() {
+        if (!global.enableSyncTests) {
+            return;
+        }
+
+        const realm = new Realm({sync: true});
+        TestCase.assertThrowsContaining(() =>  {
+            realm.privileges();
+        }, 'Wrong Realm type');
+        TestCase.assertThrowsContaining(() =>  {
+            realm.privileges('__Role');
+        }, 'Wrong Realm type');
+        TestCase.assertThrowsContaining(() =>  {
+            realm.permissions();
+        }, 'Wrong Realm type');
+        TestCase.assertThrowsContaining(() =>  {
+            realm.permissions('__Class');
+        }, 'Wrong Realm type');
+        TestCase.assertThrowsContaining(() =>  {
+            realm.subscriptions();
+        }, 'Wrong Realm type');
+        TestCase.assertThrowsContaining(() =>  {
+            realm.unsubscribe('foo');
+        }, 'Wrong Realm type');
+    } ,
+
 };